# backend/src/main/java/com/inventory/auth/application/port/in/dto/response/AuthResponse.java
package com.inventory.auth.application.port.in.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {

    private String accessToken;
    private String tokenType;
    private String email;
    private String firstName;
    private String lastName;
    private String role;
}